update balance test failing
